Fix some tidy stuff

+++ relearn/source/sim/Simulation.cpp
@@ -253,7 +253,7 @@ void Simulation::simulate(const size_t number_steps) {
             if (disable_step == step) {
                 LogFiles::write_to_file(LogFiles::EventType::Cout, true, "Disabling {} neurons in step {}", disable_ids.size(), disable_step);
                 const auto num_deleted_synapses = neurons->disable_neurons(disable_ids);
-                total_synapse_deletions += num_deleted_synapses;
+                total_synapse_deletions += static_cast<int64_t>(num_deleted_synapses);
             }
         }
